requirements.txt
absl-py==0.14.1
albumentations==1.1.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.1.5
brambox==3.2.0
cachetools==4.1.0
certifi==2021.5.30
chainer==7.7.0
chardet==3.0.4
charset-normalizer==2.0.4
chumpy==0.70
click==8.0.3
cPython==0.0.6
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
deepface==0.0.68
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
facenet-pytorch==2.5.2
fastrlock==0.5
filelock==3.0.12
filterpy==1.4.5
Flask==2.0.2
flatbuffers==1.12
fonttools==4.28.2
freetype-py==2.2.0
future==0.18.2
gast==0.4.0
gdown==4.1.1
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.4
importlib-metadata==1.6.1
insightface==0.5
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.32.1
lmdb==0.98
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.5.0
mistune==0.8.4
moviepy==1.0.3
mtcnn==0.1.1
multi-person-tracker==0.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.47.0
numpy==1.18.5
oauthlib==3.1.0
onnx==1.10.2
onnxruntime==1.9.0
opencv-python==4.2.0.34
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prettytable==2.4.0
proglog==0.1.9
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycocotools==2.0.3
pycuda==2019.1.2
pyglet==1.5.18
Pygments==2.6.1
pymongo==3.11.0
PyOpenGL==3.1.0
pyparsing==2.4.7
pypi==2.1
pyrender==0.1.36
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytools==2020.3
pytube==10.9.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qudida==0.0.4
requests==2.23.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.5.0
setuptools-scm==6.3.2
six==1.15.0
smplx==0.1.26
soupsieve==2.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
tomli==1.2.2
torch==1.9.0+cu111
torchaudio==0.9.0
torchgeometry==0.1.2
torchvision==0.10.0+cu111
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
trimesh==3.5.25
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
yacs==0.1.6
yolov3==0.1
zipp==3.1.0
